+++ src/components/pages/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import NavBar from "../reusable-ui/NavBar";
+import NavBar from "../reusable-ui/Navbar/NavBar";
 
 export default function Home() {
   return (
