+++ src/components/reusable-ui/Logo.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Logo() {
+  return (
+    <div>
+      <img src="/images/logo.png" alt="Logo de kasa" />
+    </div>
+  );
+}
